*** empty log message ***
git-svn-id: https://ceph.svn.sf.net/svnroot/ceph@508 29311d96-e01e-0410-9327-a35deaab8ce9

+++ ceph/client/Buffercache.cc
@@ -19,7 +19,7 @@ Bufferhead::Bufferhead(Inode *inode, Buffercache *bc) :
   // buffers are allocated later
 }
 
-Bufferhead::Bufferhead(Inode *inode, long long off, Buffercache *bc) : 
+Bufferhead::Bufferhead(Inode *inode, off_t off, Buffercache *bc) : 
   ref(0), miss_len(0), dirty_since(0), visited(false) {
   dout(10) << "bc: new bufferhead ino: " << inode->ino() << " offset: " << off << endl;
   this->inode = inode;
@@ -59,21 +59,21 @@ Bufferhead::~Bufferhead()
   }   
 }
 
-void Bufferhead::set_offset(long long offset)
+void Bufferhead::set_offset(off_t offset)
 {
   this->offset = offset;
   assert(!fc->buffer_map.count(offset)); // fail loudly if offset already exists!
   fc->insert(offset, this); 
 }
 
-void Bufferhead::alloc_buffers(unsigned long long size)
+void Bufferhead::alloc_buffers(off_t size)
 {
   dout(10) << "bc: allocating buffers size: " << size << endl;
   assert(size > 0);
   while (size > 0) {
     if (size <= (unsigned)g_conf.client_bcache_alloc_maxsize) {
-          unsigned long long k = g_conf.client_bcache_alloc_minsize;
-          unsigned long long asize = size - size % k + (size % k > 0) * k;
+          off_t k = g_conf.client_bcache_alloc_minsize;
+          off_t asize = size - size % k + (size % k > 0) * k;
 	  buffer *b = new buffer(asize);
 	  b->set_length(size);
 	  bl.push_back(b);
@@ -92,7 +92,7 @@ void Bufferhead::alloc_buffers(unsigned long long size)
   assert(bl.length() == size);
 }
 
-void Bufferhead::miss_start(unsigned long long miss_len) 
+void Bufferhead::miss_start(off_t miss_len) 
 {
   assert(state == BUFHD_STATE_CLEAN);
   get();
@@ -195,7 +195,7 @@ void Bufferhead::claim_append(Bufferhead *other)
   other->bl.clear();
 }
 
-void Bufferhead::splice(long long rel_off, unsigned long long length, Bufferhead *claim_by)
+void Bufferhead::splice(off_t rel_off, off_t length, Bufferhead *claim_by)
 {
   dout(10) << "bc: Bufferhead::splice rel_off: " << rel_off << " length: " << length << " claim_by: " << claim_by << endl;
   assert(length <= this->length());
@@ -280,7 +280,7 @@ bool Dirtybuffers::exist(Bufferhead* bh)
 }
 
 
-void Dirtybuffers::get_expired(time_t ttl, unsigned long long left_dirty, set<Bufferhead*>& to_flush) 
+void Dirtybuffers::get_expired(time_t ttl, off_t left_dirty, set<Bufferhead*>& to_flush) 
 {
   dout(6) << "bc: get_expired ttl: " << ttl << " left_dirty: " << left_dirty << endl;
   if (_dbufs.empty() || left_dirty >= bc->get_dirty_size()) {
@@ -288,8 +288,8 @@ void Dirtybuffers::get_expired(time_t ttl, unsigned long long left_dirty, set<Bu
     assert(_dbufs.size() == _revind.size());
     return;
   }
-  map<Inode*, map<long long, list<long long> > > consolidation_map;
-  unsigned long long cleaned = 0;
+  map<Inode*, map<off_t, list<off_t> > > consolidation_map;
+  off_t cleaned = 0;
   if (cleaned < bc->get_dirty_size() - left_dirty) {
     time_t now = time(NULL);
     for (multimap<time_t, Bufferhead*>::iterator it = _dbufs.begin();
@@ -307,15 +307,15 @@ void Dirtybuffers::get_expired(time_t ttl, unsigned long long left_dirty, set<Bu
 	  assert(!fc->buffer_map.empty());
 	}
 	assert(fc->buffer_map.count(it->second->offset));
-	list<long long> offlist;
-	unsigned long long length = fc->consolidation_opp(
+	list<off_t> offlist;
+	off_t length = fc->consolidation_opp(
 			  now - ttl, 
 			  bc->get_dirty_size() - left_dirty - cleaned, 
 			  it->second->offset, 
 			  offlist);
 	if (offlist.size() > 1) {
 	  offlist.sort();
-	  long long start_off = offlist.front();
+	  off_t start_off = offlist.front();
 	  offlist.pop_front();
 	  consolidation_map[it->second->inode][start_off] = offlist;
 	  to_flush.insert(fc->buffer_map[start_off]);
@@ -336,15 +336,15 @@ void Dirtybuffers::get_expired(time_t ttl, unsigned long long left_dirty, set<Bu
 
 // -- Filecache methods
 
-void Filecache::insert(long long offset, Bufferhead* bh)
+void Filecache::insert(off_t offset, Bufferhead* bh)
 {
-  pair<map<long long, Bufferhead*>::iterator, bool> rvalue;
-  rvalue = buffer_map.insert(pair<long long, Bufferhead*> (offset, bh));
+  pair<map<off_t, Bufferhead*>::iterator, bool> rvalue;
+  rvalue = buffer_map.insert(pair<off_t, Bufferhead*> (offset, bh));
 
   // The following is just to get the pieces for the last two assertions 
-  map<long long, Bufferhead*>::iterator next_buf = buffer_map.upper_bound(offset);
+  map<off_t, Bufferhead*>::iterator next_buf = buffer_map.upper_bound(offset);
 
-  map<long long, Bufferhead*>::iterator prev_buf = rvalue.first;
+  map<off_t, Bufferhead*>::iterator prev_buf = rvalue.first;
   if (prev_buf != buffer_map.begin()) {
     prev_buf--;
   } else {
@@ -358,7 +358,7 @@ void Filecache::insert(long long offset, Bufferhead* bh)
          (unsigned)prev_buf->first + prev_buf->second->length() <= (unsigned)offset);
 }
 
-void Filecache::splice(long long offset, unsigned long long size)
+void Filecache::splice(off_t offset, off_t size)
 {
   // insert Bufferhead at offset with size. only works if all overlapping
   // buffers are clean. Creates at most two new bufferheads at (offset, size)
@@ -371,7 +371,7 @@ void Filecache::splice(long long offset, unsigned long long size)
   // FIXME: does not work with sparse files
 #if 0
   dout(1) << "bc: before align offset: " << offset << " size: " << size << endl;
-  unsigned long long align = g_conf.client_bcache_align;
+  off_t align = g_conf.client_bcache_align;
   while (inode->inode.layout.stripe_size % align) align >>= 1;
   offset = offset / align * align;
   size = (size / align + (size % align > 0)) * align;
@@ -379,9 +379,9 @@ void Filecache::splice(long long offset, unsigned long long size)
 #endif
   
   // get current buffer
-  map<long long, Bufferhead*>::iterator curbuf = get_buf(offset);
+  map<off_t, Bufferhead*>::iterator curbuf = get_buf(offset);
   assert(curbuf != buffer_map.end());
-  unsigned long long orig_len = curbuf->second->length();
+  off_t orig_len = curbuf->second->length();
 
   // insert new buffer leaving front part to original buffer
   if (curbuf->second->state == BUFHD_STATE_CLEAN && curbuf->first < offset ) {
@@ -432,9 +432,9 @@ void Filecache::splice(long long offset, unsigned long long size)
 }
 
 
-map<long long, Bufferhead*>::iterator Filecache::get_buf(long long off)
+map<off_t, Bufferhead*>::iterator Filecache::get_buf(off_t off)
 {
-  map<long long, Bufferhead*>::iterator curbuf = buffer_map.lower_bound(off);
+  map<off_t, Bufferhead*>::iterator curbuf = buffer_map.lower_bound(off);
   if (curbuf == buffer_map.end() || curbuf->first > off) {
     if (curbuf == buffer_map.begin()) {
       return buffer_map.end();
@@ -451,7 +451,7 @@ map<long long, Bufferhead*>::iterator Filecache::get_buf(long long off)
   }
 }
 
-map<long long, Bufferhead*>::iterator Filecache::overlap(unsigned long long len, long long off)
+map<off_t, Bufferhead*>::iterator Filecache::overlap(off_t len, off_t off)
 {
   // returns iterator to buffer overlapping specified extent or end() if no overlap exists
   dout(7) << "bc: overlap " << len << " " << off << endl;
@@ -459,7 +459,7 @@ map<long long, Bufferhead*>::iterator Filecache::overlap(unsigned long long len,
   if (buffer_map.empty()) return buffer_map.end();
 
   // find first buffer with offset >= off
-  map<long long, Bufferhead*>::iterator it = buffer_map.lower_bound(off);
+  map<off_t, Bufferhead*>::iterator it = buffer_map.lower_bound(off);
 
   // Found buffer with exact offset
   if (it != buffer_map.end() && it->first == off) {
@@ -490,18 +490,18 @@ map<long long, Bufferhead*>::iterator Filecache::overlap(unsigned long long len,
   return buffer_map.end();
 }
 
-map<long long, Bufferhead*>::iterator 
-Filecache::map_existing(unsigned long long len, 
-			long long start_off,
-			map<long long, Bufferhead*>& hits, 
-			map<long long, Bufferhead*>& rx, 
-			map<long long, Bufferhead*>& tx, 
-			map<long long, unsigned long long>& holes)
+map<off_t, Bufferhead*>::iterator 
+Filecache::map_existing(off_t len, 
+			off_t start_off,
+			map<off_t, Bufferhead*>& hits, 
+			map<off_t, Bufferhead*>& rx, 
+			map<off_t, Bufferhead*>& tx, 
+			map<off_t, off_t>& holes)
 {
   dout(7) << "bc: map_existing len: " << len << " off: " << start_off << endl;
-  long long need_off = start_off;
-  long long actual_off = start_off;
-  map<long long, Bufferhead*>::iterator existing, rvalue = overlap(len, start_off);
+  off_t need_off = start_off;
+  off_t actual_off = start_off;
+  map<off_t, Bufferhead*>::iterator existing, rvalue = overlap(len, start_off);
   for (existing = rvalue;
        existing != buffer_map.end() && (unsigned)existing->first < (unsigned)start_off + len;
        existing++) {
@@ -511,7 +511,7 @@ Filecache::map_existing(unsigned long long len,
 
     if (actual_off > need_off) {
       assert(buffer_map.count(need_off) == 0);
-      holes[need_off] = (unsigned long long) (actual_off - need_off);
+      holes[need_off] = actual_off - need_off;
       dout(6) << "bc: map: hole " << need_off << " " << holes[need_off] << endl;
       need_off = actual_off;
     } 
@@ -542,27 +542,27 @@ Filecache::map_existing(unsigned long long len,
 
   // no buffers or no buffers at tail
   if ((unsigned)need_off < (unsigned)start_off + len) {
-    holes[need_off] = (unsigned long long) (start_off + len - need_off);
+    holes[need_off] = start_off + len - need_off;
     dout(6) << "bc: map: last hole " << need_off << " " << holes[need_off] << endl;
     assert(buffer_map.count(need_off) == 0);
   }
   return rvalue;
 }
 
-unsigned long long 
-Filecache::consolidation_opp(time_t max_dirty_since, unsigned long long clean_goal, 
-			     long long offset, list<long long>& offlist)
+off_t 
+Filecache::consolidation_opp(time_t max_dirty_since, off_t clean_goal, 
+			     off_t offset, list<off_t>& offlist)
 {
   dout(6) << "bc: consolidation_opp max_dirty_since: " << max_dirty_since << " clean_goal: " << clean_goal << " offset: " << offset << endl;
-  unsigned long long length = 0;
-  map<long long, Bufferhead*>::iterator cur, orig = buffer_map.find(offset);
+  off_t length = 0;
+  map<off_t, Bufferhead*>::iterator cur, orig = buffer_map.find(offset);
   assert(orig != buffer_map.end());
   length += orig->second->length();
   offlist.push_back(offset);
 
   // search left
   cur = orig;
-  long long need_off = offset;
+  off_t need_off = offset;
   while (cur != buffer_map.begin()) {
     cur--;
     if (cur->second->state != BUFHD_STATE_DIRTY  ||
@@ -597,17 +597,17 @@ Filecache::consolidation_opp(time_t max_dirty_since, unsigned long long clean_go
 void Filecache::get_dirty(set<Bufferhead*>& to_flush)
 {
   dout(6) << "bc: fc.get_dirty" << endl;
-  map<Inode*, map<long long, list<long long> > > consolidation_map;
+  map<Inode*, map<off_t, list<off_t> > > consolidation_map;
   for (set<Bufferhead*>::iterator it = dirty_buffers.begin();
        it != dirty_buffers.end();
        it++) {
     if (!(*it)->visited) {
       (*it)->visited = true;
-      list<long long> offlist;
+      list<off_t> offlist;
       consolidation_opp( time(NULL), bc->get_dirty_size(), (*it)->offset, offlist);
       if (offlist.size() > 1) {
         offlist.sort();
-	long long start_off = offlist.front();
+	off_t start_off = offlist.front();
 	consolidation_map[inode][start_off] = offlist;
 	to_flush.insert(buffer_map[start_off]);
       } else {
@@ -626,7 +626,7 @@ void Filecache::simplify()
 {
   dout(7) << "bc: simplify" << endl;
   list<Bufferhead*> removed;
-  map<long long, Bufferhead*>::iterator start, next;
+  map<off_t, Bufferhead*>::iterator start, next;
   start = buffer_map.begin();
   next = buffer_map.begin();
   int count = 0;
@@ -669,14 +669,14 @@ void Filecache::simplify()
 }
 #endif
 
-int Filecache::copy_out(unsigned long long size, long long offset, char *dst) 
+int Filecache::copy_out(off_t size, off_t offset, char *dst) 
 {
   dout(7) << "bc: copy_out size: " << size << " offset: " << offset << endl;
   assert(offset >= 0);
   //assert(offset + size <= length()); doesn't hold after trim_bcache
   int rvalue = size;
   
-  map<long long, Bufferhead*>::iterator curbuf = buffer_map.lower_bound(offset);
+  map<off_t, Bufferhead*>::iterator curbuf = buffer_map.lower_bound(offset);
   if (curbuf == buffer_map.end() || curbuf->first > offset) {
     if (curbuf == buffer_map.begin()) {
       return -1;
@@ -717,14 +717,14 @@ int Filecache::copy_out(unsigned long long size, long long offset, char *dst)
 
 // -- Buffercache methods
 
-void Buffercache::dirty(Inode *inode, unsigned long long size, long long offset, const char *src) 
+void Buffercache::dirty(Inode *inode, off_t size, off_t offset, const char *src) 
 {
   dout(6) << "bc: dirty ino: " << inode->ino() << " size: " << size << " offset: " << offset << endl;
   assert(bcache_map.count(inode->ino())); // filecache has to be already allocated!!
   Filecache *fc = get_fc(inode);
   assert(offset >= 0);
   
-  map<long long, Bufferhead*>::iterator curbuf = fc->get_buf(offset);
+  map<off_t, Bufferhead*>::iterator curbuf = fc->get_buf(offset);
   assert(curbuf != fc->buffer_map.end());
 
   if (curbuf->second->state == BUFHD_STATE_CLEAN) {
@@ -753,17 +753,17 @@ void Buffercache::dirty(Inode *inode, unsigned long long size, long long offset,
   }
 }
 
-unsigned long long Buffercache::touch_continuous(map<long long, Bufferhead*>& hits, unsigned long long size, long long offset)
+off_t Buffercache::touch_continuous(map<off_t, Bufferhead*>& hits, off_t size, off_t offset)
 {
   dout(7) << "bc: touch_continuous size: " << size << " offset: " << offset << endl;
   if (hits.empty()) return 0;
 
-  long long next_off = offset;
+  off_t next_off = offset;
   if (hits.begin()->first > offset ||
       (unsigned)hits.begin()->first + hits.begin()->second->length() <= (unsigned)offset) {
     return 0;
   }
-  for (map<long long, Bufferhead*>::iterator curbuf = hits.begin(); 
+  for (map<off_t, Bufferhead*>::iterator curbuf = hits.begin(); 
        curbuf != hits.end();
        curbuf++) {
     if (curbuf == hits.begin()) {
@@ -774,21 +774,21 @@ unsigned long long Buffercache::touch_continuous(map<long long, Bufferhead*>& hi
     lru.lru_touch(curbuf->second);
     next_off += curbuf->second->length();
   }
-  return (unsigned long long)(next_off - offset) >= size ? size : (next_off - offset);
+  return (next_off - offset) >= size ? size : (next_off - offset);
 }
 
-void Buffercache::map_or_alloc(Inode *inode, unsigned long long size, long long offset, 
-                               map<long long, Bufferhead*>& buffers, 
-                               map<long long, Bufferhead*>& rx,
-			       map<long long, Bufferhead*>& tx)
+void Buffercache::map_or_alloc(Inode *inode, off_t size, off_t offset, 
+                               map<off_t, Bufferhead*>& buffers, 
+                               map<off_t, Bufferhead*>& rx,
+			       map<off_t, Bufferhead*>& tx)
 {
   dout(7) << "bc: map_or_alloc len: " << size << " off: " << offset << endl;
   Filecache *fc = get_fc(inode);
-  map<long long, unsigned long long> holes;
+  map<off_t, off_t> holes;
   holes.clear();
   fc->map_existing(size, offset, buffers, rx, tx, holes);
   // stuff buffers into holes
-  for (map<long long, unsigned long long>::iterator hole = holes.begin();
+  for (map<off_t, off_t>::iterator hole = holes.begin();
        hole != holes.end();
        hole++) {
     Bufferhead *bh;
@@ -807,19 +807,19 @@ void Buffercache::map_or_alloc(Inode *inode, unsigned long long size, long long 
   }
 }
 
-void Buffercache::consolidate(map<Inode*, map<long long, list<long long> > > cons_map)
+void Buffercache::consolidate(map<Inode*, map<off_t, list<off_t> > > cons_map)
 {
   dout(6) << "bc: consolidate" << endl;
   int deleted = 0;
-  for (map<Inode*, map<long long, list<long long> > >::iterator it_ino = cons_map.begin();
+  for (map<Inode*, map<off_t, list<off_t> > >::iterator it_ino = cons_map.begin();
        it_ino != cons_map.end();
        it_ino++) {
     Filecache *fc = get_fc(it_ino->first);
-    for (map<long long, list<long long> >::iterator it_off = it_ino->second.begin();
+    for (map<off_t, list<off_t> >::iterator it_off = it_ino->second.begin();
          it_off != it_ino->second.end();
 	 it_off++) {
       Bufferhead *first_bh = fc->buffer_map[it_off->first];
-      for (list<long long>::iterator it_list = it_off->second.begin();
+      for (list<off_t>::iterator it_list = it_off->second.begin();
            it_list != it_off->second.end();
 	   it_list++) {
 	Bufferhead *bh = fc->buffer_map[*it_list];
@@ -837,7 +837,7 @@ void Buffercache::consolidate(map<Inode*, map<long long, list<long long> > > con
   dout(6) << "bc: consolidate: deleted: " << deleted << endl;
 }
 
-void Buffercache::get_reclaimable(unsigned long long min_size, list<Bufferhead*>& reclaimed)
+void Buffercache::get_reclaimable(off_t min_size, list<Bufferhead*>& reclaimed)
 {
   while (min_size > 0) {
     if (Bufferhead *bh = (Bufferhead*)lru.lru_expire()) {
@@ -850,10 +850,10 @@ void Buffercache::get_reclaimable(unsigned long long min_size, list<Bufferhead*>
 }
 
 
-unsigned long long Buffercache::reclaim(unsigned long long min_size)
+off_t Buffercache::reclaim(off_t min_size)
 {
   dout(7) << "bc: reclaim min_size: " << min_size << endl;
-  unsigned long long freed_size = 0;
+  off_t freed_size = 0;
   while (freed_size < min_size) {
     Bufferhead *bh = (Bufferhead*)lru.lru_expire();
     if (!bh) {
